+++ client/gui/CGuiHandler.cpp
@@ -615,7 +615,7 @@ void CFramerateManager::framerateDelay()
 	// FPS is higher than it should be, then wait some time
 	if (timeElapsed < rateticks)
 	{
-		SDL_Delay(ceil(this->rateticks) - timeElapsed);
+		SDL_Delay((Uint32)ceil(this->rateticks) - timeElapsed);
 	}
 
 	accumulatedTime += timeElapsed;
@@ -624,7 +624,7 @@ void CFramerateManager::framerateDelay()
 	if(accumulatedFrames >= 100)
 	{
 		//about 2 second should be passed
-		fps = ceil(1000.0 / (accumulatedTime/accumulatedFrames));
+		fps = static_cast<int>(ceil(1000.0 / (accumulatedTime/accumulatedFrames)));
 		accumulatedTime = 0;
 		accumulatedFrames = 0;
 	}
